* commit 'f557803a90a29967a9fa41f043e4fd304b1aeee5':
  libsel4debug: Accept untyped sizes up to word size
  libsel4debug: Fix debug printing
  libsel4debug: Remove 32-bit size assumptions
  libsel4debug: Define x86 registers by sel4_arch instead of arch
  libsel4simple-default: Fix 32-bit type assumptions in printing
  libsel4simple-default: Remove 32-bit size assumptions
  libsel4simple: Remove 32-bit size assumptions
  libsel4allocman: Update x86_64 object creation
  libsel4allocman: Fix 32/64bit compatiblities in LOG functions

+++ libsel4simple-default/src/libsel4simple-default.c
@@ -13,6 +13,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <assert.h>
+#include <inttypes.h>
 
 #include <sel4/sel4.h>
 #include <sel4debug/debug.h>
@@ -123,7 +124,7 @@ int simple_default_untyped_count(void *data) {
     return ((seL4_BootInfo *)data)->untyped.end - ((seL4_BootInfo *)data)->untyped.start;
 }
 
-seL4_CPtr simple_default_nth_untyped(void *data, int n, uint32_t *size_bits, uint32_t *paddr, bool *device) {
+seL4_CPtr simple_default_nth_untyped(void *data, int n, size_t *size_bits, uintptr_t *paddr, bool *device) {
     assert(data && size_bits && paddr);
 
     seL4_BootInfo *bi = data;
